backend/requirements.txt
flask
python-dotenv
flask-restful
flask-cors
openai
langchain